refactor(runs): extract artifact preview/capture + runner-attach into core
The runs artifact command owned two orchestration clusters that tripped
the thin_command_adapter audit detector (direct process execution +
filesystem mutation):

1. Static-artifact preview/screenshot capture: preview and capture each
   inlined their own python3 http.server spawn (literal duplication),
   plus the sh-based capture provider, port reservation, entrypoint
   resolution, and manifest write.
2. Runner artifact-attach source acquisition: local/SSH source resolve,
   remote download to a temp dir, and temp cleanup.

Extract two focused core services:
- core::observation::artifact_preview: start_static_artifact_server (now
  shared by preview and capture, killing the duplication),
  capture_artifact_entrypoints engine, provider resolution, manifest
  write, and entrypoint/port helpers.
- core::observation::runner_artifact_attach: RunnerAttachSource types,
  copy_runner_artifact_source, and cleanup_runner_attach_source.

The command now maps args -> core call -> output types and delegates all
process/filesystem work. Command module drops ~486 lines. The capture
entrypoint-safety unit test moves to the core module with its code.

Clears the runs/remote_artifact.rs thin_command_adapter_violation.
