+++ src/core/observation/mod.rs
@@ -4,6 +4,7 @@
 //! stack specs, baselines). SQLite stores observed state from command runs and
 //! generated artifacts. This module only provides the storage substrate.
 
+pub mod artifact_preview;
 mod budget_findings;
 pub mod bundle;
 pub mod context;
@@ -14,6 +15,7 @@ pub mod loop_inventory_run;
 pub mod observed_workflow;
 pub mod records;
 mod run_failure_causes;
+pub mod runner_artifact_attach;
 pub mod runs_service;
 pub mod store;
 mod test_findings;
